+++ client/src/pages/Notice/components/NoticeListTable/NoticeRow/NoticeRow.module.css
@@ -1,5 +1,5 @@
 .row td {
-  padding: 12px 8px;
+  padding: 1.2rem 0.8rem;
   vertical-align: middle;
 }
