--- client/src/pages/Notice/components/NoticeSection/Pagination/Pagination.jsx
@@ -1,83 +1,0 @@
-import { mockNotice } from '../../../mock/mockNotice';
-import styles from './Pagination.module.css';
-
-
-export default function Pagination({ currentPage, onPageChange, activeTab }) {
-  const itemsPerPage = 10;
-  const filtered = mockNotice.filter((item) => item.category === activeTab);
-  const totalPages = Math.ceil(filtered.length / itemsPerPage);
-
-  const pageNumbers = getPageNumbers(currentPage, totalPages, 5);
-
-  return (
-    <div className={styles.paginationContainer}>
-      {currentPage > 1 && (
-        <button
-          onClick={() => onPageChange(1)}
-          className={styles.pageButton}
-        >
-          «
-        </button>
-      )}
-
-      {currentPage > 1 && (
-        <button
-          onClick={() => onPageChange(currentPage - 1)}
-          className={styles.pageButton}
-        >
-          ‹
-        </button>
-      )}
-
-      {pageNumbers.map((page) => (
-        <button
-          key={page}
-          onClick={() => onPageChange(page)}
-          className={`${styles.pageButton} ${currentPage === page ? styles.active : ''}`}
-        >
-          {page}
-        </button>
-      ))}
-
-      {currentPage < totalPages && (
-        <button
-          onClick={() => onPageChange(currentPage + 1)}
-          className={styles.pageButton}
-        >
-          ›
-        </button>
-      )}
-
-      {currentPage < totalPages && (
-        <button
-          onClick={() => onPageChange(totalPages)}
-          className={styles.pageButton}
-        >
-          »
-        </button>
-      )}
-    </div>
-
-  );
-}
-
-function getPageNumbers(currentPage, totalPages, maxVisible = 5) {
-  const half = Math.floor(maxVisible / 2);
-  let start = Math.max(1, currentPage - half);
-  let end = Math.min(totalPages, currentPage + half);
-
-  if (end - start + 1 < maxVisible) {
-    if (start === 1) {
-      end = Math.min(totalPages, start + maxVisible - 1);
-    } else if (end === totalPages) {
-      start = Math.max(1, end - maxVisible + 1);
-    }
-  }
-
-  const result = [];
-  for (let i = start; i <= end; i++) {
-    result.push(i);
-  }
-  return result;
-}
-
