+++ client/src/pages/Notice/mock/mockNotice.js
@@ -1,5 +1,3 @@
-// src/pages/Notice/data/mockNotice.js
-
 export const mockNotice = [
   {
     id: 1,
