+++ client/src/pages/NoticeDetail/components/ContentsCard/ContentsCard.jsx
