+++ client/src/pages/NoticeDetail/components/ContentsDetailSection/ContentsDetailSection.jsx
@@ -1,13 +1,15 @@
-import React from 'react';
 import styles from './ContentsDetailSection.module.css';
 
 export default function ContentsDetailSection({ content }) {
   return (
     <div className={styles.contentsDetail}>
       <h2 className={styles.heading}>공모내용</h2>
-      {content.split('\n').map((line, idx) =>
-        line.trim() ? <p key={idx}>{line}</p> : <br key={idx} />
-      )}
+      {content.split('\n').map((line) => {
+        const trimmed = line.trim();
+        if (!trimmed) return <br key={crypto.randomUUID()} />;
+        return <p key={trimmed}>{trimmed}</p>;
+      })}
     </div>
   );
 }
+
