--- client/src/pages/NoticeDetail/components/NoticeContentsSection/NoticeContentsSection.module.css
@@ -1,35 +1,0 @@
-.noticeLayout {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  padding: 2rem;
-  margin: 0 auto;
-  gap: 2rem;
-}
-
-.contentsContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-
-  width: 100%;
-  gap: 2rem;
-}
-
-.divider {
-  width: 100%;
-  
-  border: none;
-  border-top: 0.1rem solid #e5e5e5;
-}
-
-
-.infoWrapper {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-
-  max-width: 100%;
-  gap: 2rem;
-}
